requirements.txt
# Wiki.js MCP Server Dependencies
# Python 3.12+ required

# Core MCP Framework
fastmcp>=0.1.0

# Web Server for ChatGPT Integration
fastapi>=0.104.0
uvicorn>=0.24.0

# HTTP Client
httpx>=0.27.0

# Data Validation and Settings
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Text Processing
python-slugify>=8.0.0
markdown>=3.5.0
beautifulsoup4>=4.12.0

# Environment variable loading
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Retry logic
tenacity>=8.0.0

# Development Dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0 